sync
git-svn-id: https://daisy-model.googlecode.com/svn/trunk@954 5f5bf4bb-3343-0410-8278-4fa8eb3319dc

+++ time.C
@@ -160,7 +160,7 @@ Time::tick_day (int days)
 
 void
 Time::tick_year (int years)
-{ impl.year += years; }
+{ impl.year += static_cast<short> (years); }
 
 // @ Convert.
